Add/remove domain users

+++ src/com/telifie/Models/Clients/DomainsClient.java
@@ -3,9 +3,9 @@ package com.telifie.Models.Clients;
 import com.mongodb.MongoException;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
-import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.telifie.Models.Domain;
+import com.telifie.Models.Member;
 import com.telifie.Models.Utilities.Configuration;
 import org.bson.Document;
 import java.util.ArrayList;
@@ -14,9 +14,6 @@ import java.util.List;
 
 public class DomainsClient extends Client {
 
-    /**
-     *
-     */
     public DomainsClient(Configuration config){
         super(config);
         this.collection = "domains";
@@ -27,30 +24,21 @@ public class DomainsClient extends Client {
      * @param userId ID of user
      * @return ArrayList<Domain>
      */
-    public ArrayList<Domain> getOwnedDomains(String userId){
-
+    public ArrayList<Domain> ownedDomains(String userId){
         ArrayList<Document> found = super.find(new Document("owner", userId));
         ArrayList<Domain> domains = new ArrayList<>();
         for(Document doc : found){
             domains.add(new Domain(doc));
         }
-
         return domains;
     }
 
-    public Domain getWithId(String id){
+    public Domain withId(String id){
         return new Domain(super.findOne(new Document("id", id)));
     }
 
     public boolean delete(String owner, String id){
-        return super.deleteOne(
-            new Document("$and",
-                Arrays.asList(
-                    new Document("owner", owner),
-                    new Document("id", id)
-                )
-            )
-        );
+        return super.deleteOne(new Document("$and", Arrays.asList(new Document("owner", owner), new Document("id", id))));
     }
 
     /**
@@ -59,37 +47,32 @@ public class DomainsClient extends Client {
      * @param userId ID of user
      * @return ArrayList<Domain>
      */
-    public ArrayList<Domain> getDomainsForUser(String userId){
-
+    public ArrayList<Domain> forUser(String userId){
         ArrayList<Domain> domains = new ArrayList<>();
-        super.find(
-                new Document("$and", List.of(
-                        new Document("owner", userId)
-                )
-            )
-        );
-
+        super.find(new Document("$and", List.of(new Document("owner", userId))));
         return domains;
     }
 
     public boolean create(Domain domain){
-
         if(this.insertOne(Document.parse(domain.toString()))){
-
-            //Now setup collections in the databases to be prepared for use.
             try(MongoClient mongoClient = MongoClients.create(super.config.getDomain().getUri())){
-
                 MongoDatabase database = mongoClient.getDatabase("domains-articles");
                 database.createCollection(domain.getAlt());
-
             }catch(MongoException e){
-
                 return true;
             }
-
             return true;
-
         }
         return false;
+    }
+
+    public boolean addUsers(String domain, ArrayList<Member> members){
+        members.forEach(member -> super.updateOne(new Document("id", domain), new Document("$push", new Document("users", Document.parse(member.toString())))));
+        return true;
+    }
+
+    public boolean removeUsers(String domain, ArrayList<Member> members){
+        members.forEach(member -> super.updateOne(new Document("id", domain), new Document("$pull", new Document("users", Document.parse(member.toString())))));
+        return true;
     }
 }
